- fix warning related to amd64
git-svn-id: https://openser.svn.sourceforge.net/svnroot/openser/trunk@4894 689a6050-402a-0410-94f2-e92a70836424

+++ modules_k/db_postgres/res.c
@@ -116,7 +116,7 @@ int db_postgres_get_columns(const db_con_t* _h, db_res_t* _r)
 			db_free_columns(_r);
 			return -4;
 		}
-		LM_DBG("allocate %d bytes for RES_NAMES[%d] at %p\n", sizeof(str), col,
+		LM_DBG("allocate %d bytes for RES_NAMES[%d] at %p\n", (unsigned int) sizeof(str), col,
 				RES_NAMES(_r)[col]);
 
 		/* The pointer that is here returned is part of the result structure. */
